[ddc] Run co19 nnbd tests in weak mode
Adjusted the shard count for the "_2" suites running in weak mode
from 6 to 5. Gave the co19 tests 2 shards since there are far fewer
tests in that suite right now.

Compared to another from earlier today the total execution time
appears to be roughly the same except for the deflaking of the
new and failing tests.

Change-Id: I71122af8ba84138e8044a57550302f3b37b6671b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/147821
Reviewed-by: Riley Porter <rileyporter@google.com>
Commit-Queue: Nicholas Shahan <nshahan@google.com>