[dart2js] Reduce Lists in DeferredLoadTask
ImportSets do not need to store the list of members of the set as it
is available implicitly on the canonical construction path. All that
is required is a back-pointer to traverse the canonical path from the
end (i.e. ImportSet).

This reduces the memory required by DeferredLoadTask by O(N).

Change-Id: Id23ce91d4aa01a74cd0f9ff4246dc56b65db1bf9
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/148184
Commit-Queue: Stephen Adams <sra@google.com>
Reviewed-by: Sigmund Cherem <sigmund@google.com>