Collapse package:js changelog
We haven't published any of the recent changes to `package:js`, revert
back to the version just after the latest published.

Remove the changelog entry referring to `promiseToFuture`. This was a change in
the SDK libraries and is not associated with any single version of `package:js`.

Remove the author field from the pubspec, it is unused.

Expand the pubspec description.

Change-Id: Id2a37a7c947656a6f5ff4dd577d0e7d32e8e0b61
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/148188
Auto-Submit: Nate Bosch <nbosch@google.com>
Reviewed-by: Kevin Moore <kevmoo@google.com>
Commit-Queue: Kevin Moore <kevmoo@google.com>