[dart2js] treat native return types as nullable in weak mode.
This is a temporary workaround. The `dart:html` library contains type
annotations based on the chrome IDL, but it doesn't consider browser
compatibility or whether an API is only supported on some contexts.

We can revert this as soon as `dart:html` is sound again.

See https://github.com/dart-lang/sdk/issues/41960 for details.

Change-Id: Ia92e8821adc8e3fab8cd8e08327a282c133962ca
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/148622
Reviewed-by: Stephen Adams <sra@google.com>
Commit-Queue: Sigmund Cherem <sigmund@google.com>