Reland [nnbd_migration] Fix two popup style bugs
Relands https://dart-review.googlesource.com/c/sdk/+/148642

Due to a z-index bug, regions are showing up on top of the popup. Fixed.

A regression during a previous a.button/.button change that meant that
the close button wasn't rendered as a button.

Previous change from `a.button/.button` reduced the specificity which
caused incorrect font color in edit link buttons. This has been fixed by
using `.button, a.button` so that we keep the desired specificity but
also keep applying the style to all elements with the `button` class.

Change-Id: Idb5eb154904430d946a80b7574537564b7c71bd1
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/148646
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Mike Fairhurst <mfairhurst@google.com>