[infra] Improve triggering for analyzer trybots.
Previously, analyzer-analysis-server-linux and analyzer-linux-release
were triggered on all CLs.  They really only need to be triggered on
changes to pkg/ (as well as the set of standard paths that trigger
everything, such as tests).

Previously, analyzer-nnbd-linux-release wasn't triggered at all.  It
needs to be triggered on changes to the analyzer itself (and standard
paths).

Change-Id: I56581724b572aac41926aa7d7bad03c444bff2a0
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/150141
Reviewed-by: Alexander Thomas <athom@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>