Analyzer: report conflicting extension type variables
Fixes https://github.com/dart-lang/sdk/issues/42216

Report when an extension declares a type parameter with the same name as the
extension.

Report when an extension declares a member with the same basename as the name
of any of the extension's type parameters.

Move tests into diagnostics/.

Move tests for RECURSIVE_INTERFANCE_INHERITANCE.

Change-Id: If19041571008b9b14122d9ff458d1d30ea0ff034
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/150270
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>