Add common terminals to ANSI heuristic
Today, Stdout and Stderr have a property supportsAnsiEscapes. The
documented purpose of this property is to inform the caller whether ANSI
escape sequences are supported - say, for colorizing terminal output
when the user is running the Dart program in a terminal emulator like
xterm.

The problem is that the heuristic used is a substring match for "xterm"
in the TERM environment variable. (See attached bug.) This
implementation omits a few significant populations of users:

* Users using screen or tmux. These commands set their TERM to "screen".
* Users using a non-xterm terminal like rxvt (especially rxvt-unicode).

The correct solution here is to query terminfo for terminal
capabilities. Until then, adding these two cases to the existing xterm
check unblocks users of two popular terminal multiplexers and a terminal
emulator.

Bug: https://github.com/dart-lang/sdk/issues/31606
Change-Id: I1a8a37d42f1895d7b1238c67ec86207fd4a38bbb
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/149940
Auto-Submit: Wesley Moy <wesleymoy@google.com>
Reviewed-by: Jonas Termansen <sortie@google.com>
Reviewed-by: Zichang Guo <zichangguo@google.com>
Commit-Queue: Jonas Termansen <sortie@google.com>