[vm/aot] Add Image objects to V8 snapshot profiles.
This avoids attributing the header and padding bytes in Image objects
to the artificial root, instead making it clear why these bytes exist.
It also ensures Instructions objects (in the non-bare-payload case) are
never considered roots of the profile.

Also unify semantics of text_offset in AssemblyImageWriter::WriteText()
and BlobImageWriter::WriteText(). Originally the former was the offset
from the start of the text section and the latter was the offset from
the start of the instructions objects/payloads. Now they're both the
offset in the text section.

Bug: https://github.com/dart-lang/sdk/issues/41249

Cq-Include-Trybots: luci.dart.try:vm-kernel-precomp-linux-release-x64-try,vm-kernel-precomp-linux-product-x64-try,vm-kernel-precomp-linux-debug-x64-try,vm-kernel-precomp-linux-release-simarm_x64-try
Change-Id: I617e76c9c390a93abb6bc37efc1355368c64691e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/149286
Reviewed-by: Vyacheslav Egorov <vegorov@google.com>
Commit-Queue: Tess Strickland <sstrickl@google.com>